move from pinentry-qt4 to pinentry-qt
we're also providing a transitional package for the old pinentry-qt4
name -- upstream is clearly intending to switch to the new
pinentry-qt.

I've reduced the default priority for pinentry-qt in
update-alternatives from 95 to 80, since upstream seems to be focusing
on GNOME integration with a higher priority than KDE at the moment.